// packages/control-property-editor/src/api.ts
export type PropertyValue = string | boolean | number;

export interface PropertyChange<T extends PropertyValue = PropertyValue> {
    controlId: string;
    controlName: string;
    propertyName: string;
    value: T;
}